Update DayTrader3 Dashboard to DayTrader7 Dashboard related to #118,#223 Signed-off-by: Dong Hoon Lee <l.donghoon49@gmail.com>
