+++ test-result-summary-client/src/Dashboard/Widgets/Graph/index.js
@@ -1,4 +1,4 @@
-export { default as DayTrader3 } from "./DayTrader3";
+export { default as DayTrader7 } from "./DayTrader7";
 export { default as ODM } from "./ODM";
 export { default as SPECjbb2015 } from "./SPECjbb2015";
 export { default as AcmeairNodejs} from "./AcmeairNodejs";
